+++ R/learner_randomForestSRC_surv_rfsrc.R
@@ -131,7 +131,8 @@ LearnerSurvRandomForestSRC = R6Class("LearnerSurvRandomForestSRC",
         cores = p_int(default = 1L, lower = 1L, tags = c("train", "predict", "threads")),
         save.memory = p_lgl(default = FALSE, tags = "train"),
         perf.type = p_fct(levels = "none", tags = "train"),
-        case.depth = p_lgl(default = FALSE, tags = c("train", "predict"))
+        case.depth = p_lgl(default = FALSE, tags = c("train", "predict")),
+        marginal.xvar = p_uty(default = NULL, tags = "predict")
       )
 
       ps$values$ntime = 0
@@ -187,10 +188,7 @@ LearnerSurvRandomForestSRC = R6Class("LearnerSurvRandomForestSRC",
       pv = convert_ratio(pv, "sampsize", "sampsize.ratio", task$nrow)
       cores = pv$cores %??% 1L
       pv$cores = NULL
-
-      if ("weights" %in% task$properties) {
-        pv$case.wt = as.numeric(task$weights$weight) # nolint
-      }
+      pv$case.wt = private$.get_weights(task) # nolint
 
       invoke(randomForestSRC::rfsrc,
         formula = task$formula(), data = task$data(),